app/alloy-agent/pyproject.toml
[project]
name = "naonous-agent"
version = "1.0.0"
description = "A2A agent for managing BYOK campaign lifecycle on Alloy"
requires-python = ">=3.12"
dependencies = [
    # Core A2A
    "a2a-sdk[http-server]>=0.3.9",
    "uvicorn>=0.38.0",
    "click>=8.1.8",
    # DeepAgents and LangChain
    "deepagents>=0.1.4,<0.2.0",
    "langchain>=1.0.0",
    "langchain-aws>=1.1.0",
    "langchain-mcp-adapters>=0.1.3",
    # Checkpointing
    "langgraph>=0.3.1",
    "langgraph-checkpoint-aws>=1.0.1",
    "aiodynamo>=24.7",
    # HTTP clients
    "httpx>=0.28.1",
    # Data models
    "pydantic>=2.10.6",
    # Auth
    "pyjwt>=2.10.1",
    "authlib>=1.6.5",
    # Shared SDKs
    "rcplus-alloy-common>=8.6.1",
    "ringier-a2a-sdk>=0.1.0",
    # Utilities
    "python-dotenv>=1.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.6",
    "mypy>=1.15.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Allow unused imports in __init__.py

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "--cov=agent --cov-report=html --cov-report=term-missing"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.uv.sources]
ringier-a2a-sdk = { path = "../ringier-a2a-sdk", editable = true }
